+++ examples/showoci/showoci.py
@@ -47,10 +47,20 @@
 # - oci.healthchecks.HealthChecksClient
 # - oci.announcements_service.AnnouncementClient
 # - oci.limits.LimitsClient
+# - oci.integration.IntegrationInstanceClient
+# - oci.analytics.AnalyticsClient
+# - oci.oda.OdaClient
+# - oci.oce.OceInstanceClient
+# - oci.apigateway.GatewaysClient
+# - oci.functions.FunctionsManagementClient
 #
 # Modules Not Yet Covered:
 # - oci.waas.WaasClient
 # - oci.dns.DnsClient
+# - oci.data_catalog.DataCatalogClient
+# - oci.data_flow.DataFlowClient
+# - oci.data_science.DataScienceClient
+# - oci.events.EventsClient
 #
 ##########################################################################
 from __future__ import print_function
@@ -62,7 +72,7 @@ import sys
 import argparse
 import datetime
 
-version = "20.1.29"
+version = "20.2.11"
 
 ##########################################################################
 # execute_extract
@@ -219,24 +229,27 @@ def set_parser_arguments():
     parser.add_argument('-a', action='store_true', default=False, dest='all', help='Print All Resources')
     parser.add_argument('-ani', action='store_true', default=False, dest='allnoiam', help='Print All Resources but identity')
     parser.add_argument('-an', action='store_true', default=False, dest='announcement', help='Print Announcements')
+    parser.add_argument('-api', action='store_true', default=False, dest='api', help='Print API Gateways')
     parser.add_argument('-b', action='store_true', default=False, dest='budgets', help='Print Budgets')
-    parser.add_argument('-n', action='store_true', default=False, dest='network', help='Print Network')
-    parser.add_argument('-i', action='store_true', default=False, dest='identity', help='Print Identity')
-    parser.add_argument('-ic', action='store_true', default=False, dest='identity_compartments', help='Print Identity Compartments only')
     parser.add_argument('-c', action='store_true', default=False, dest='compute', help='Print Compute')
     parser.add_argument('-cn', action='store_true', default=False, dest='container', help='Print Containers')
-    parser.add_argument('-o', action='store_true', default=False, dest='object', help='Print Object Storage')
-    parser.add_argument('-l', action='store_true', default=False, dest='load', help='Print Load Balancer')
     parser.add_argument('-d', action='store_true', default=False, dest='database', help='Print Database')
-    parser.add_argument('-f', action='store_true', default=False, dest='file', help='Print File Storage')
     parser.add_argument('-e', action='store_true', default=False, dest='email', help='Print EMail')
-    parser.add_argument('-m', action='store_true', default=False, dest='monitoring', help='Print Monitoring and Notifications')
-    parser.add_argument('-s', action='store_true', default=False, dest='streams', help='Print Streams')
-    parser.add_argument('-rm', action='store_true', default=False, dest='orm', help='Print Resource management')
-    parser.add_argument('-so', action='store_true', default=False, dest='sumonly', help='Print Summary Only')
-    parser.add_argument('-paas', action='store_true', default=False, dest='paas_native', help='Print Oracle Paas Native - OIC OAC ODA')
     parser.add_argument('-edge', action='store_true', default=False, dest='edge', help='Print Edge Services (Healthcheck)')
+    parser.add_argument('-f', action='store_true', default=False, dest='file', help='Print File Storage')
+    parser.add_argument('-fun', action='store_true', default=False, dest='function', help='Print Functions')
+    parser.add_argument('-i', action='store_true', default=False, dest='identity', help='Print Identity')
+    parser.add_argument('-ic', action='store_true', default=False, dest='identity_compartments', help='Print Identity Compartments only')
+    parser.add_argument('-l', action='store_true', default=False, dest='load', help='Print Load Balancer')
     parser.add_argument('-lq', action='store_true', default=False, dest='limits', help='Print Limits and Quotas')
+    parser.add_argument('-m', action='store_true', default=False, dest='monitoring', help='Print Monitoring and Notifications')
+    parser.add_argument('-n', action='store_true', default=False, dest='network', help='Print Network')
+    parser.add_argument('-o', action='store_true', default=False, dest='object', help='Print Object Storage')
+    parser.add_argument('-paas', action='store_true', default=False, dest='paas_native', help='Print Oracle Paas Native - OIC OAC ODA')
+    parser.add_argument('-rm', action='store_true', default=False, dest='orm', help='Print Resource management')
+    parser.add_argument('-s', action='store_true', default=False, dest='streams', help='Print Streams')
+
+    parser.add_argument('-so', action='store_true', default=False, dest='sumonly', help='Print Summary Only')
     parser.add_argument('-mc', action='store_true', default=False, dest='mgdcompart', help='exclude ManagedCompartmentForPaaS')
     parser.add_argument('-nr', action='store_true', default=False, dest='noroot', help='Not include root compartment')
     parser.add_argument('-ip', action='store_true', default=False, dest='instance_principals', help='Use Instance Principals for Authentication')
@@ -244,7 +257,7 @@ def set_parser_arguments():
     parser.add_argument('-p', default="", dest='proxy', help='Set Proxy (i.e. www-proxy-server.com:80) ')
     parser.add_argument('-rg', default="", dest='region', help='Filter by Region')
     parser.add_argument('-cp', default="", dest='compart', help='Filter by Compartment Name or OCID')
-    parser.add_argument('-cpr', default="", dest='compart_recursive', help='Filter by Compartment Name or OCID include sub compartments')
+    parser.add_argument('-cpr', default="", dest='compart_recur', help='Filter by Comp Name or OCID Recursive')
     parser.add_argument('-cpath', default="", dest='compartpath', help='Filter by Compartment path ,(i.e. -cpath "Adi / Sub"')
     parser.add_argument('-tenantid', default="", dest='tenantid', help='Override confile file tenancy_id')
     parser.add_argument('-cf', type=argparse.FileType('r'), dest='config', help="Config File")
@@ -265,7 +278,8 @@ def set_parser_arguments():
     if not (result.all or result.allnoiam or result.network or result.identity or result.identity_compartments or
             result.compute or result.object or
             result.load or result.database or result.file or result.email or result.orm or result.container or
-            result.streams or result.budgets or result.monitoring or result.edge or result.announcement or result.limits or result.paas_native):
+            result.streams or result.budgets or result.monitoring or result.edge or result.announcement or result.limits or result.paas_native or
+            result.api or result.function):
 
         parser.print_help()
 
@@ -328,6 +342,12 @@ def set_service_extract_flags(cmd):
     if cmd.all or cmd.allnoiam or cmd.budgets:
         prm.read_budgets = True
 
+    if cmd.all or cmd.allnoiam or cmd.function:
+        prm.read_function = True
+
+    if cmd.all or cmd.allnoiam or cmd.api:
+        prm.read_api = True
+
     if cmd.all or cmd.allnoiam or cmd.limits:
         prm.read_limits = True
 
@@ -359,8 +379,8 @@ def set_service_extract_flags(cmd):
     if cmd.compart:
         prm.filter_by_compartment = str(cmd.compart)
 
-    if cmd.compart_recursive:
-        prm.filter_by_compartment_recursive = str(cmd.compart_recursive)
+    if cmd.compart_recur:
+        prm.filter_by_compartment_recursive = str(cmd.compart_recur)
 
     if cmd.compartpath:
         prm.filter_by_compartment_path = str(cmd.compartpath)